+++ Public.DTO/v1/Isik.cs
@@ -15,7 +15,7 @@ public class Isik
     [MaxLength(64)]
     public string Perenimi { get; set; }
     
-    public int Isikukood { get; set; }
+    public long Isikukood { get; set; }
     
     [MaxLength(1500)]
     public string? Lisainfo { get; set; }
